+++ Client.Linq/Internal/Expressions/IExpressionPart.cs
@@ -5,7 +5,7 @@ namespace InfluxDB.Client.Linq.Internal.Expressions
     internal interface IExpressionPart
     {
         /// <summary>
-        /// Append Flux Query to buiilder
+        /// Append Flux Query to builder.
         /// </summary>
         /// <param name="builder">Flux query builder</param>
         void AppendFlux(StringBuilder builder);