+++ Client.Linq/Internal/Expressions/LeftParenthesis.cs
@@ -1,6 +1,6 @@
 namespace InfluxDB.Client.Linq.Internal.Expressions
 {
-    internal class LeftParenthesis: AbstractExpressionPart
+    internal class LeftParenthesis : AbstractExpressionPart
     {
         internal LeftParenthesis() : base("(")
         {